Move code around a bit
